add movideails from api

+++ src/main/java/com/example/MyMovies/dtos/AddRatingDto.java
@@ -2,5 +2,5 @@ package com.example.MyMovies.dtos;
 
 import java.util.UUID;
 
-public record AddRatingDto(UUID userId, String comment, int rating, int tmdbId) {
+public record AddRatingDto(UUID userId, String comment, int rating, long tmdbId) {
 }
